requirements.txt
streamlit==1.11.1
pandas
plotly_express==0.4.0
firebase_admin